SBA-94 Added Directory Layout Action Buttons
Added Action Buttons for Directory Layout - Save, Clone and Delete.
Create, Update, Delete of Directory Layout are done. The callbacks to update the parent component's state is yet to be implemented. So for now, a manual reload of the page is needed.

Added Bonus - Accordion and Buttons are implemented using material-ui