SBA-94 Fixed bug in Banner Selection.
Cancelling the Banner Selection after successful upload resulted in error. Possibly due to the HTML file field being uncontrolled.

Mostly likely cause is React clears up the files selected during the state change, but somehow the internal state of the field (maintained by browser's JS Engine) remains unchanged (file selected is still 1 in the internal state), and when file selection is cancelled, it detects there is a change in the files selected (previously 1, now 0), thus triggering the onChange handler.

Thus a guard condition is added to continue only if the files are actually selected.